requirements.txt
openai
streamlit
python-dotenv
cohere
chromadb
